pyproject.toml
[project]
name = "research_assistant_challenge"
version = "0.1.0"
description = "code containing research assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4",
    "langchain-text-splitters>=1.0.0",
    "langchain[openai]>=1.0.2",
    "langgraph>=1.0.1",
    "letta-client==0.1.324",
    "markitdown[all]>=0.1.3",
    "nltk>=3.9.2",
    "onnxruntime>=1.15.1",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.12.0",
    "unstructured[md]>=0.18.15",
]
